split components into their own go module
