+++ components/baggage_span_processor_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package honeycomb
+package components
 
 import (
 	"context"
