+++ components/deterministic_sampler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package honeycomb
+package components
 
 import (
 	"go.opentelemetry.io/otel/attribute"
