+++ examples/baggage/example.go
@@ -25,8 +25,8 @@ import (
 	"time"
 
 	"github.com/honeycombio/honeycomb-opentelemetry-go"
+	"github.com/honeycombio/honeycomb-opentelemetry-go/components"
 	"github.com/honeycombio/otel-config-go/otelconfig"
-
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/baggage"
@@ -39,7 +39,7 @@ func main() {
 			attribute.Int64("app.unix_time_ms", time.Now().UnixMilli()),
 		}
 	})
-	bsp := honeycomb.NewBaggageSpanProcessor()
+	bsp := components.NewBaggageSpanProcessor()
 
 	shutdown, err := otelconfig.ConfigureOpenTelemetry(
 		otelconfig.WithSpanProcessor(dsp, bsp),
